stash1

+++ src/main/java/com/ey/bank/realtime_transaction_tracker/dao/TransactionRequest.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 public class TransactionRequest {
     private String transactionId;
     private String sender;
-    private double amount;
+    private Integer amount;
     private String receiver;
     private String status;
     // Getters and setters
